% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbGetResults.R
\name{addDefaultValues}
\alias{addDefaultValues}
\title{addDefaultValues
If default values are not overwritten OpenML does not return values. This function adds values for the tuned hyperparameters so they can be used for the calculation of the pareto front.}
\usage{
addDefaultValues(res)
}
\arguments{
\item{res}{Long table with hyperparameters generated in getHyperparTable}
}
\value{
[\code{data.frame}] Long Table with added values for the defaults.
}

