% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paretoBenchmarkParetoFront.R
\name{benchmarkParetoFront}
\alias{benchmarkParetoFront}
\title{benchmarkParetoFront
Predicts the paretofront for an out-of-bag dataset and compares it to the real paretofront of the dataset.}
\usage{
benchmarkParetoFront(measure.name = "area.under.roc.curve", learner.name,
  task.ids, lrn.par.set, tbl.results, tbl.hypPars, tbl.metaFeatures,
  tbl.runTime, tbl.resultsReference, surrogate.mlr.lrn, min.experiments = 100,
  n.points1 = 10000, n.points2 = 1000)
}
\arguments{
\item{measure.name}{}

\item{learner.name}{}

\item{task.ids}{}

\item{lrn.par.set}{}

\item{tbl.results}{}

\item{tbl.hypPars}{}

\item{tbl.metaFeatures}{}

\item{tbl.runTime}{}

\item{tbl.resultsReference}{}

\item{surrogate.mlr.lrn}{mlr.learner used for creating the surrogate function}

\item{min.experiments}{take experiments with a least min.experiments}

\item{n.points1}{Number of points for first paretofront (prediction)}

\item{n.points2}{Number of points for the real paretofront of the dataset}
}
