% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paretoCreateParetoFront.R
\name{createParetoFront}
\alias{createParetoFront}
\title{createParetoFront}
\usage{
createParetoFront(learner.name, lrn.par.set, surrogates.measures,
  surrogates.time, meta.features, n.points = 100, extra.points = NULL)
}
\arguments{
\item{learner.name}{}

\item{lrn.par.set}{}

\item{surrogates.measures}{}

\item{surrogates.time}{}

\item{meta.features}{}

\item{n.points}{}

\item{extra.points}{}
}
\description{
Create a pareto front for a given surrogate model for a measure and time of a learner
}
