% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paretoCreateSurrogateModel.R
\name{deleteNA}
\alias{deleteNA}
\title{Replace NA values by -11 (numeric variables) or NA levels (factor variables)}
\usage{
deleteNA(task.data)
}
\arguments{
\item{task.data}{task data}
}
\description{
Replace NA values by -11 (numeric variables) or NA levels (factor variables)
}
