% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/botEvaluationRun.R
\name{evalConfigurations}
\alias{evalConfigurations}
\title{evalConfigurations
This evaluates all configurations of a single learner and a matching task.
Configurations in par should be valid for task.}
\usage{
evalConfigurations(lrn, task, par, min.resources, max.resources, upload, path)
}
\arguments{
\item{lrn}{Learner}

\item{task}{OMLTask}

\item{par}{data.frame of configurations to evaluate}

\item{min.resources}{minimal used resources}

\item{max.resources}{maximum allowed resources for a single evaluation}

\item{upload}{should the run be uploaded}
}
