% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/botexponentialBackOff.R
\name{exponentialBackOff}
\alias{exponentialBackOff}
\title{exponentialBackOff}
\usage{
exponentialBackOff(jobs, registry, start.resources = NULL, max.resources)
}
\arguments{
\item{jobs}{job numbers}

\item{registry}{used batchtools registry}

\item{start.resources}{starting resources, if not specified use the default resources}

\item{max.resources}{maximum allowed resources. This should be a list with slots walltime and memory.}
}
\description{
Gradually increase the maximum walltime and memory requirements
}

