% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbGetResults.R
\name{getHyperparTable}
\alias{getHyperparTable}
\title{getHyperparTable}
\usage{
getHyperparTable(run.tag = "mlrRandomBot", numRuns = 4e+05,
  excl.run.ids = NULL, local.db = NULL, n = 4e+05)
}
\arguments{
\item{run.tag}{}

\item{numRuns}{maximum Number of runs that should be downloaded in the first step; should be set to a value so that it downloads all available runs}

\item{excl.run.ids}{ids that should not be downloaded}

\item{local.db}{}

\item{n}{maximum number of runs that should be downloaded}
}
\value{
[\code{data.frame}] Table with run.id, hyperparameter name & value.
}
\description{
get the table with hyperparameters from OpenML
}

