% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/botSetLearnerParamPairs.R
\name{getMultipleLearners}
\alias{getMultipleLearners}
\title{getMultipleLearners}
\usage{
getMultipleLearners()
}
\value{
a learner parameter set
}
\description{
Expand the simple learner set by more learners.
}
