% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbGetResults.R
\name{getReferenceTable}
\alias{getReferenceTable}
\title{getReferenceTable
contains the results for the featureless learner used for scaling}
\usage{
getReferenceTable(run.tag = "referenceV1", numRuns = 320000,
  excl.run.ids = NULL, local.db = NULL)
}
\arguments{
\item{run.tag}{}

\item{numRuns}{}

\item{excl.run.ids}{}

\item{local.db}{}
}
\value{
[\code{data.frame}] Table with run.id, task.id, flow.id, flow.name, measure values.
}

