% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResults.R
\name{getRunDf}
\alias{getRunDf}
\title{Small helper function to get a clean df from listOMLRuns}
\usage{
getRunDf(run.tag, numRuns, excl.run.ids)
}
\description{
Small helper function to get a clean df from listOMLRuns
}

