% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbGetResults.R
\name{getRunTable}
\alias{getRunTable}
\title{getRunTable
get the run table from OpenML and melt the measure values}
\usage{
getRunTable(run.tag = "mlrRandomBot", numRuns = 320000,
  excl.run.ids = NULL, local.db = NULL)
}
\arguments{
\item{run.tag}{}

\item{numRuns}{}

\item{excl.run.ids}{}

\item{local.db}{}
}
\value{
[\code{data.frame}] Table with run.id, task.id, flow.id, flow.name, measure values.
}
