% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbGetResults.R
\name{getRunTimeTable}
\alias{getRunTimeTable}
\title{getRunTimeTable
Get runtime table from the database}
\usage{
getRunTimeTable(local.db = NULL, numRuns = 320000)
}
\arguments{
\item{local.db}{}

\item{numRuns}{}
}
