% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/botSetLearnerParamPairs.R
\name{getSimpleLearners}
\alias{getSimpleLearners}
\title{getSimpleLearners}
\usage{
getSimpleLearners()
}
\value{
a learner parameter set
}
\description{
Define some simple learners with parameter sets to be explored by the bot
}
