% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paretoCreateParetoFront.R
\name{getTimeDependentHyperpar}
\alias{getTimeDependentHyperpar}
\title{getTimeDependentHyperpar
Get a proposal for a hyperparameter setting for a given time}
\usage{
getTimeDependentHyperpar(surrogate_measures, surrogate_time, meta_features)
}
\arguments{
\item{surrogate_measures}{}

\item{surrogate_time}{Name of learner}

\item{meta_features}{}
}
