% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databaseManagement.R
\name{initializeLocalDatabase}
\alias{initializeLocalDatabase}
\title{initializeLocalDatabase}
\usage{
initializeLocalDatabase(path = ".", overwrite = FALSE)
}
\arguments{
\item{path}{path of database}

\item{overwrite}{overwrite existing database}
}
\value{
database
}
\description{
Initialize database
}

