% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/botMakeLrnPsSets.R
\name{makeLrnPsSets}
\alias{makeLrnPsSets}
\title{makeLrnPsSets}
\usage{
makeLrnPsSets(learner, param.set, lrn.ps.sets = NULL,
  id = paste0(learner$id, ".set"), overwrite = FALSE)
}
\arguments{
\item{learner}{learner object}

\item{param.set}{parameter set, that matches the learner}

\item{lrn.ps.sets}{a previously created lrn.ps.set}

\item{id}{name of the lrn.ps set to create}

\item{overwrite}{overwrite existing id, if id collision occurs}
}
\description{
Create a learner-parameter set.
}

