% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paretoSurrogateModel.R
\name{makeMeasureTimePrediction}
\alias{makeMeasureTimePrediction}
\title{Create random hyperpars for a flow and predict the measure and time for a given task.}
\usage{
makeMeasureTimePrediction(measure.name, learner.name, task.id, lrn.par.set, n,
  tbl.results, tbl.hypPars, tbl.metaFeatures)
}
\arguments{
\item{measure.name}{Name of the measure to optimize}

\item{learner.name}{Name of learner}

\item{task.id}{ID of the task}

\item{lrn.par.set}{learner-parameter set which should include relevant bounds for flow}

\item{n}{Number of points to create}

\item{tbl.results}{df with getMlrRandomBotResults()}

\item{tbl.hypPars}{df with getMlrRandomBotHyperpars()}

\item{tbl.metaFeatures}{df with getMlrRandomBotHyperpars()}
}
\value{
[\code{data.frame}] df with hyperpars of flow and matching predictions for measure value and time
}
\description{
Create random hyperpars for a flow and predict the measure and time for a given task.
}

