% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paretoSurrogateModel.R
\name{makeRandomTable}
\alias{makeRandomTable}
\title{Create table for a specific task with random hyperpars}
\usage{
makeRandomTable(task.id.filter, learner.name, lrn.ps.sets, n, tbl.metaFeatures)
}
\arguments{
\item{task.id.filter}{What task to use}

\item{learner.name}{What learner to use}

\item{lrn.ps.sets}{What learner to use}

\item{n}{Number of rows to create}

\item{tbl.hypPars}{df with getMlrRandomBotHyperpars()}
}
\value{
[\code{data.frame}] Complete table used for exploring surrogate model
}
\description{
Create table for a specific task with random hyperpars
}

