% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paretoCreateSurrogateModel.R
\name{makeSurrogateModel}
\alias{makeSurrogateModel}
\title{makeSurrogateModel
Create surrogate model for different tasks}
\usage{
makeSurrogateModel(measure.name, learner.name, task.ids, lrn.par.set,
  tbl.results, tbl.hypPars, tbl.metaFeatures, tbl.runTime, tbl.resultsReference,
  surrogate.mlr.lrn, min.experiments = 100, benchmark = FALSE,
  time = FALSE)
}
\arguments{
\item{measure.name}{Name of the measure to optimize}

\item{learner.name}{Name of learner}

\item{task.ids}{[\code{numeric}] ids of the tasks}

\item{lrn.par.set}{learner-parameter set which should include relevant bounds for flow}

\item{tbl.results}{df with getMlrRandomBotResults()}

\item{tbl.hypPars}{df with getMlrRandomBotHyperpars()}

\item{tbl.metaFeatures}{df with getMlrRandomBotHyperpars()}

\item{tbl.runTime}{}

\item{tbl.resultsReference}{}

\item{surrogate.mlr.lrn}{model(s) that should be used as surrogate model}

\item{min.experiments}{minimum number of experiments that should be available for a dataset, otherwise the dataset is excluded}

\item{benchmark}{[\code{logical}] Should a benchmark experiment with different surrogate models be executed to}

\item{time}{evaluate the performance of the surrogate models? Default is FALSE.}
}
\value{
surrogate model
}

