% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paretoCreateParetoFront.R
\name{plotParetoFront}
\alias{plotParetoFront}
\title{plotParetoFront
Plot the pareto front that was created with createParetoFront}
\usage{
plotParetoFront(learner.name, par.front, plotly = FALSE, log = FALSE,
  col = NULL, cex = NULL)
}
\arguments{
\item{learner.name}{}

\item{par.front}{pareto front object created with createParetoFront}

\item{plotly}{should plotly be used? Default is FALSE.}

\item{log}{should the time axis be scaled on a log scale? Default is FALSE.}

\item{col}{vector of colors for extra hyperparameters if they are available in par.front. Default is NULL.}

\item{cex}{vector of point size for extra hyperparameters if they are available in par.front. Default is NULL.}
}
