% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/botCallWrapper.R
\name{runBot}
\alias{runBot}
\title{runBot
Main function to run the bot. This evaluates batch.size configurations. 
Task, learner and the actual configuration are defined by the corresponding functions.}
\usage{
runBot(batch.size, sample.learner.fun = sampleRandomLearner,
  sample.task.fun = sampleSimpleTask,
  sample.configuration.fun = sampleRandomConfiguration,
  min.resources = NULL, max.resources = NULL,
  lrn.ps.sets = simple.lrn.par.set, upload = FALSE, path = NA,
  extra.tag = "botV1")
}
\arguments{
\item{batch.size}{number of configurations to evaluate in parallel}

\item{sample.learner.fun}{function to sample a learner from tunePair}

\item{sample.task.fun}{sample a OML task}

\item{min.resources}{minimal resources to start benchmark (list with elements walltime and memory)}

\item{max.resources}{maximum resources allowed for each evaluation (list with elements walltime and memory)}

\item{lrn.ps.sets}{of available learners with matching parameter sets}

\item{upload}{should the run be uploaded to OpenML}

\item{path}{where should the registry be created, defaults to tempdir()}

\item{sample.configuration}{given a lrn sample a configuration}
}
