% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/botSampleConfigs.R
\name{sampleRandomAzureTask}
\alias{sampleRandomAzureTask}
\title{sampleRandomAzureTask}
\usage{
sampleRandomAzureTask(bot.nr)
}
\arguments{
\item{bot.nr}{number of the bot that is runnning}
}
\value{
OML task
}
\description{
this draws a random binary classif OMLTask from study 14 with 10 fold CV and without missing values
}

