% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/botSampleConfigs.R
\name{sampleRandomConfiguration}
\alias{sampleRandomConfiguration}
\title{Samples size random Configuration for a given parameter set}
\usage{
sampleRandomConfiguration(size, par.set)
}
\arguments{
\item{size}{number of configurations to generate}

\item{par.set}{parameter set}
}
\value{
data.frame where each row is one valid configuration
}
\description{
Samples size random Configuration for a given parameter set
}

