% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/botSampleConfigs.R
\name{sampleRandomLearner}
\alias{sampleRandomLearner}
\title{Sample a random learner with matching parameter set from the lrn.ps.sets list}
\usage{
sampleRandomLearner(lrn.ps.sets)
}
\arguments{
\item{lrn.ps.sets}{of available learners with matching parameter sets}
}
\value{
list of one learner with matching parameter set
}
\description{
Sample a random learner with matching parameter set from the lrn.ps.sets list
}
