% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/botSampleConfigs.R
\name{sampleRandomTask}
\alias{sampleRandomTask}
\title{sampleRandomTask}
\usage{
sampleRandomTask()
}
\value{
OML task
}
\description{
this draws a random binary classif OMLTask from study 14 with 10 fold CV and without missing values
}
