% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paretoCreateSurrogateModel.R
\name{scaleRunTime}
\alias{scaleRunTime}
\title{Scale the run time results with the sci.mark results}
\usage{
scaleRunTime(tbl.runTime)
}
\arguments{
\item{tbl.runTime}{The runtime table}
}
\description{
Scale the run time results with the sci.mark results
}
