% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResults.R
\name{scrapeRunTime}
\alias{scrapeRunTime}
\title{Scrape user time and scibench for a run. Only works for runs by OMLBot}
\usage{
scrapeRunTime(run.id)
}
\description{
Scrape user time and scibench for a run. Only works for runs by OMLBot
}

