% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/botSetLearnerParamPairs.R
\name{testlearner}
\alias{testlearner}
\title{testlearner
Just for testing one learner}
\usage{
testlearner()
}
\value{
learner param set
}
\description{
testlearner
Just for testing one learner
}

