% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databaseManagement.R
\name{updateHyperparTable}
\alias{updateHyperparTable}
\title{updateHyperparTable}
\usage{
updateHyperparTable(run.tag, local.db)
}
\arguments{
\item{run.tag}{}

\item{local.db}{}
}
\description{
Save hyperparameters for run to db
}

