% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbManagement.R
\name{updateLocalDatabase}
\alias{updateLocalDatabase}
\title{updateLocalDatabase}
\usage{
updateLocalDatabase(path = ".", run.tag = "mlrRandomBot",
  task.tag = "study_14")
}
\arguments{
\item{path}{path of database}

\item{run.tag}{tag for OMLRun}

\item{task.tag}{tag for OMLTask}
}
\value{
database
}
\description{
Check each tables of the local database for update requirement
}
