% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databaseManagement.R
\name{updateMetaTable}
\alias{updateMetaTable}
\title{updateMetaTable}
\usage{
updateMetaTable(task.tag, local.db)
}
\arguments{
\item{task.tag}{}

\item{local.db}{}
}
\description{
Save meta data for task to db
}

