% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databaseManagement.R
\name{updateReferenceTable}
\alias{updateReferenceTable}
\title{updateReferenceTable}
\usage{
updateReferenceTable(run.tag, local.db)
}
\arguments{
\item{run.tag}{}

\item{local.db}{}
}
\description{
Save reference run results to db
}

