% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbGetResults.R
\name{getRunDf}
\alias{getRunDf}
\title{getRunDf
Small helper function to get a clean df from listOMLRuns}
\usage{
getRunDf(run.tag, numRuns, excl.run.ids)
}
\arguments{
\item{run.tag}{OpenML run tag}

\item{numRuns}{number of runs}

\item{excl.run.ids}{run ids to exclude}
}
