% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paretoCreateSurrogateModel.R
\name{makeBotTable}
\alias{makeBotTable}
\title{makeBotTable
Merge results, hyperpars and features tables and prepare for mlr.task input}
\usage{
makeBotTable(measure.name, learner.name, tbl.results, tbl.hypPars,
  tbl.metaFeatures, tbl.runTime, tbl.resultsReference)
}
\arguments{
\item{measure.name}{What measure to analyse}

\item{learner.name}{What learner to analyse}

\item{tbl.results}{df with getMlrRandomBotResults()}

\item{tbl.hypPars}{df with getMlrRandomBotHyperpars()}

\item{tbl.metaFeatures}{df with getMlrRandomBotHyperpars()}

\item{tbl.runTime}{}

\item{tbl.resultsReference}{}
}
\value{
[\code{data.frame}] Complete table used for creating the surrogate model
}

