% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/botSampleConfigurations.R
\name{sampleDefaultConfiguration}
\alias{sampleDefaultConfiguration}
\title{sampleDefaultConfiguration}
\usage{
sampleDefaultConfiguration(size, par.set)
}
\arguments{
\item{size}{number of configurations to generate}

\item{par.set}{parameter set}
}
\value{
data.frame where each row is one valid configuration
}
\description{
sampleDefaultConfiguration
}

