% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbGetResults.R
\name{scrapeRunTime}
\alias{scrapeRunTime}
\title{scrapeRunTime}
\usage{
scrapeRunTime(run.id)
}
\arguments{
\item{run.id}{}
}
\description{
Scrape user time and scibench for a run. Only works for runs by OMLBot
}
