% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbManagement.R
\name{updateHyperparTable}
\alias{updateHyperparTable}
\title{updateHyperparTable}
\usage{
updateHyperparTable(run.tag, local.db)
}
\arguments{
\item{run.tag}{tag of run}

\item{local.db}{local db object}
}
\description{
Save hyperparameters for run to db
}

