% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbManagement.R
\name{updateMetaTable}
\alias{updateMetaTable}
\title{updateMetaTable}
\usage{
updateMetaTable(task.tag, local.db)
}
\arguments{
\item{task.tag}{tag of task}

\item{local.db}{local db object}
}
\description{
Save meta data for task to db
}

