% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbManagement.R
\name{updateReferenceTable}
\alias{updateReferenceTable}
\title{updateReferenceTable}
\usage{
updateReferenceTable(run.tag, local.db)
}
\arguments{
\item{run.tag}{tag of run}

\item{local.db}{local db object}
}
\description{
Save reference run results to db
}

